I used this code to compare the performance:

```python
import pymc3 as pm
import numpy as np

N = 10000

with pm.Model() as model1:
    pm.Uniform('unif', lower=0, upper=1, shape=N)
    step1 = pm.step_methods.hmc.base_hmc.BaseHMC(
        scaling=np.ones(N), profile=True, use_single_leapfrog=True)

with pm.Model() as model2:
    pm.HalfNormal('unif', sd=1, shape=N)
    step2 = pm.step_methods.hmc.base_hmc.BaseHMC(
        scaling=np.ones(N), profile=True, use_single_leapfrog=True)

np.random.seed(42)
q = np.random.randn(N)
p = np.random.randn(N)
epsilon = np.array(0.01)
for model, step in zip([model1, model2], [step1, step2]):
    q_grad = model.dlogp_array(q)
    %timeit _ = step.leapfrog(q, p, q_grad, epsilon)
```

before:
100 loops, best of 3: 2.27 ms per loop
1000 loops, best of 3: 989 µs per loop

after:
1000 loops, best of 3: 1.34 ms per loop
1000 loops, best of 3: 574 µs per loop

There is still some room for improvement: The implementation of the interval transform computes `sigmoid(y)` and `softmax(-y)`. Both usually compute `exp(-y)`, but theano can't figure out that it could reuse the values. If we had an Op, that computes both at the same time we might be able to save another ~200ms:

```python
>>> %timeit np.exp(q)
1000 loops, best of 3: 226 µs per loop
```