@aseyboldt 2017-03-11
@ColCarroll When the benchmark page is finished, do you think we could put some benchmarks of leapfrog functions in there? That would help to keep changes to nuts separate from other performance improvements.