@ColCarroll 2017-03-11
Would love to -- I've bookmarked this to add to the suite once it is running.  I'm on vacation for another week, but hopefully _next_ weekend I'll get some hosted benchmarks.