@ColCarroll 2017-03-11
As far as documentation, I think the proper place is in the class docs, though @fonnesbeck or @twiecki might have an opinion.  I was going to suggest not worrying about it for this PR, but looks like it is already TeX'd up...

Looks good to merge once tests pass, though