@aseyboldt 2017-03-11
Ah, right. Tex is rendered in the docs, I forgot about that. I'll put it in then.