@aseyboldt 2017-03-11
I'm getting an error when I try to build the docs:

```
>>> cd docs/source
>>> make html
Notebook error:
TemplateNotFound in notebooks/BEST.ipynb:
nbsphinx-rst
make: *** [Makefile:55: html] Error 
```
The `conf.py` mentions a dir `templates_`, but there is no such dir.