@twiecki 2017-03-13
Very cool speed-up @aseyboldt.