@twiecki 2017-03-13
I rebuilt the docs, not sure why you had problems. Maybe missing dependencies.